문제 해결

+++ app/src/main/java/android/example/cs496/ui/main/fragment4/Menu.java
@@ -10,7 +10,7 @@ public class Menu implements Serializable {
     private int votedNumber;
     private double totalScore;
 
-    public Menu(String menuName, int menuId, int votedNumber, int totalScore) {
+    public Menu(String menuName, int menuId, int votedNumber, double totalScore) {
         this.menuName =menuName;
         this.menuId = menuId;
         this.votedNumber = votedNumber;
